// types.ts
export enum UserRole {
    Admin = 'Admin',
    Engenheiro = 'Engenheiro',
    Tecnico = 'Técnico',
    DesenhistaIndustrial = 'Desenhista Industrial',
    Desenvolvedor = 'Desenvolvedor',
}

export interface User {
    id: string;
    name: string;
    email: string;
    role: UserRole;
}